Instances module for Data.V.Linear
