+++ src/Data/V/Linear.hs
@@ -1,10 +1,5 @@
-{-# OPTIONS -Wno-orphans #-}
-{-# LANGUAGE FlexibleInstances #-}
+{-# OPTIONS_GHC -Wno-dodgy-exports #-}
 {-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE UndecidableInstances #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
-
 -- | This module defines vectors of known length which can hold linear values.
 --
 -- Having a known length matters with linear types, because many common vector
@@ -46,16 +41,11 @@ module Data.V.Linear
   , iterate
   -- * Type-level utilities
   , caseNat
+  , module Data.V.Linear.Internal.Instances
   ) where
 
 import Data.V.Linear.Internal.V
-import Prelude.Linear.Internal
-import qualified Unsafe.Linear as Unsafe
-import qualified Data.Functor.Linear.Internal.Functor as Data
-import qualified Data.Functor.Linear.Internal.Applicative as Data
-import qualified Data.Functor.Linear.Internal.Traversable as Data
-import GHC.TypeLits
-import qualified Data.Vector as Vector
+import Data.V.Linear.Internal.Instances ()
 
 {- Developers Note
 
@@ -71,21 +61,4 @@ of the instances.
 Remark: ideally the instances below would be in an internal `Instances`
 module. But we haven't got around to it yet.
 -}
-
-
--- # Instances of V
--------------------------------------------------------------------------------
-
-instance Data.Functor (V n) where
-  fmap f (V xs) = V $ Unsafe.toLinear (Vector.map (\x -> f x)) xs
-
-instance KnownNat n => Data.Applicative (V n) where
-  pure a = V $ Vector.replicate (theLength @n) a
-  (V fs) <*> (V xs) = V $
-    Unsafe.toLinear2 (Vector.zipWith (\f x -> f $ x)) fs xs
-
-instance KnownNat n => Data.Traversable (V n) where
-  traverse f (V xs) =
-    (V . Unsafe.toLinear (Vector.fromListN (theLength @n))) Data.<$>
-    Data.traverse f (Unsafe.toLinear Vector.toList xs)
